# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Документация API Filmotek",
        "title": "Filmotek API",
        "contact": {
            "name": "Поддержка API",
            "url": "https://github.com/shuklarituparn",
            "email": "rtprnshukla@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/v1/actors/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Получить всех актеров с пагинацией",
                "operationId": "read-all-actors",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество элементов на странице",
                        "name": "page_size",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Поле для сортировки (по умолчанию birth_date)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Порядок сортировки (ASC или DESC, по умолчанию DESC)",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список актеров",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReadAllActorResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный размер страницы или номер страницы",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/actors/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Создать нового актера",
                "operationId": "create-actor",
                "parameters": [
                    {
                        "description": "Объект актера, который нужно создать",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateActorModel"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Актер добавлен",
                        "schema": {
                            "$ref": "#/definitions/controllers.CreateActorResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/actors/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Удалить актера",
                "operationId": "delete-actor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Актер успешно удален",
                        "schema": {
                            "$ref": "#/definitions/controllers.DeleteActorResponse"
                        }
                    },
                    "400": {
                        "description": "Требуется идентификатор актера",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось удалить актера или его связи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/actors/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Получить актера по идентификатору",
                "operationId": "read-actor-by-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Подробности актера",
                        "schema": {
                            "$ref": "#/definitions/controllers.ReadActorResponse"
                        }
                    },
                    "400": {
                        "description": "Необходим идентификатор актера",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Актер не найден",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка при получении актера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/actors/patch": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Частично обновить существующего актера",
                "operationId": "patch-actor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор актера",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Данные для частичного обновления актера",
                        "name": "patchData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateActorModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Актер успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/controllers.PatchActorResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось обновить актера или его связи",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/actors/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Актеры"
                ],
                "summary": "Обновить существующего актера",
                "operationId": "update-actor",
                "parameters": [
                    {
                        "description": "Объект актера для обновления",
                        "name": "actor",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateActorModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Актер успешно обновлен",
                        "schema": {
                            "$ref": "#/definitions/controllers.UpdateActorResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный запрос",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Не удалось обновить актера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "controllers.CreateActorResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.DeleteActorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.PatchActorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.ReadActorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Actor"
                }
            }
        },
        "controllers.ReadAllActorResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.UpdateActorResponse": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/models.Actor"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Actor": {
            "type": "object",
            "required": [
                "birth_date",
                "first_name",
                "gender",
                "last_name"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                }
            }
        },
        "models.CreateActorModel": {
            "type": "object",
            "required": [
                "birth_date",
                "first_name",
                "gender",
                "last_name"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                }
            }
        },
...
        "models.Movie": {
            "type": "object",
            "required": [
                "genre",
                "rating",
                "release_year",
                "title"
            ],
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Actor"
                    }
                },
                "description": {
                    "type": "string",
                    "maxLength": 1000,
                    "minLength": 1
                },
                "genre": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "rating": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release_year": {
                    "type": "integer"
                },
                "title": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                }
            }
        },
...
        "models.UpdateActorModel": {
            "type": "object",
            "required": [
                "birth_date",
                "first_name",
                "gender",
                "id",
                "last_name"
            ],
            "properties": {
                "birth_date": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_name": {
                    "type": "string",
                    "maxLength": 100
                },
                "movies": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Movie"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}